More changes for the sticky scroll
