+++ score-ios/Views/ScoringSummary.swift
@@ -48,53 +48,6 @@ struct ScoringSummary : View {
     }
 }
 
-struct ScoringUpdateCell : View {
-    var update: GameUpdate
-    var img: String
-    
-    var body : some View {
-        HStack {
-            if update.isCornell {
-                Image("Cornell")
-                .resizable().frame(width: 32, height: 32)
-            } else {
-                AsyncImage(url: URL(string: img)) {image in
-                    image.resizable().frame(width: 32, height: 32)
-                } placeholder: {
-                    Constants.Colors.gray_icons.frame(width: 32, height: 32)
-                }
-            }
-            
-            Spacer()
-            
-            VStack {
-                Text("\("0:00") - \(ordinalNumberString(for: update.timestamp))")
-                    .font(Constants.Fonts.regular14)
-                
-                if update.isCornell {
-                    Text("**\(update.cornellScore)** - \(update.opponentScore)")
-                        .font(Constants.Fonts.regular14)
-                } else {
-                    Text("\(update.cornellScore) - **\(update.opponentScore)**")
-                        .font(Constants.Fonts.regular14)
-                }
-            }
-            .frame(width:72)
-            
-            Spacer()
-            
-            Text(update.description)
-                .font(Constants.Fonts.regular14)
-                .frame(width: 217)
-            
-        }
-        .padding(.leading, 24)
-        .padding(.trailing, 24)
-        .padding(.top, 12)
-        .padding(.bottom, 12)
-    }
-}
-
 #Preview {
     ScoringSummary(game: Game.dummyData[0])
 }
